# [perf] experimental enhance fp8 per-tensor quant