--- a/python/sglang/srt/layers/quantization/fp8_kernel.py
+++ b/python/sglang/srt/layers/quantization/fp8_kernel.py
@@ -842,84 +842,13 @@ def w8a8_block_fp8_matmul(
 
 
 @triton.jit
-def _per_tensor_quant_fp8(
+def _per_tensor_quant_mla_fp8_stage1(
     x_ptr,
-    x_min_ptr,
-    x_max_ptr,
     x_s_ptr,
-    x_q_ptr,
-    numel,
-    eps,
-    fp8_min,
-    fp8_max,
-    BLOCK_SIZE: tl.constexpr,
-):
-    g_id = tl.program_id(0)
-    offset = tl.arange(0, BLOCK_SIZE) + g_id * BLOCK_SIZE
-    mask = offset < numel
-
-    _xmin = tl.load(x_min_ptr).to(tl.float32)
-    _xmax = tl.load(x_max_ptr).to(tl.float32)
-    _absmax = tl.maximum(tl.maximum(tl.abs(_xmin), tl.abs(_xmax)), eps)
-    x_s_inv = fp8_max / _absmax
-
-    x = tl.load(x_ptr + offset, mask=mask, other=0.0).to(tl.float32)
-    x_q = tl.clamp(x * x_s_inv, fp8_min, fp8_max).to(x_q_ptr.dtype.element_ty)
-    tl.store(x_q_ptr + offset, x_q, mask=mask)
-
-    if g_id == 0:
-        tl.store(x_s_ptr, 1.0 / x_s_inv)
-
-
-def per_tensor_quant_fp8(
-    x: torch.Tensor, eps: float = 1e-12, dtype: torch.dtype = torch.float8_e4m3fn
-) -> Tuple[torch.Tensor, torch.Tensor]:
-    """This function quantizes input values to float8 values with tensor-wise quantization."""
-    assert x.is_contiguous(), "`x` is not contiguous"
-
-    finfo = torch.finfo(dtype)
-    fp8_max = finfo.max
-    if _is_hip:
-        dtype = torch.float8_e4m3fnuz
-        fp8_max = 224.0
-
-    x_q = torch.empty_like(x, dtype=dtype)
-    x_s = x.new_empty((1,), dtype=torch.float32)
-
-    BLOCK_SIZE = 1024
-    N = x.numel()
-    grid = (triton.cdiv(N, BLOCK_SIZE),)
-
-    min_val, max_val = x.aminmax()
-    _per_tensor_quant_fp8[grid](
-        x,
-        min_val,
-        max_val,
-        x_s,
-        x_q,
-        N,
-        eps,
-        -fp8_max,
-        fp8_max,
-        BLOCK_SIZE,
-    )
-
-    return x_q, x_s
-
-
-@triton.jit
-def _per_tensor_quant_mla_fp8(
-    x_ptr,
-    x_min_ptr,
-    x_max_ptr,
-    x_s_ptr,
-    x_q_ptr,
-    num_seq,
     head_size,
     x_stride_h,
     x_stride_s,
     eps,
-    fp8_min,
     fp8_max,
     BLOCK_SIZE: tl.constexpr,
 ):
@@ -929,19 +858,39 @@ def _per_tensor_quant_mla_fp8(
     mask = offset < head_size
 
     x_ptr += head_id * x_stride_h + seq_id * x_stride_s
-    x_q_ptr += head_id * num_seq * head_size + seq_id * head_size
+    x = tl.load(x_ptr + offset, mask=mask, other=0.0).to(tl.float32)
+    _absmax = tl.maximum(tl.max(tl.abs(x)), eps)
 
-    _xmin = tl.load(x_min_ptr).to(tl.float32)
-    _xmax = tl.load(x_max_ptr).to(tl.float32)
-    _absmax = tl.maximum(tl.maximum(tl.abs(_xmin), tl.abs(_xmax)), eps)
-    x_s_inv = fp8_max / _absmax
+    tl.atomic_max(x_s_ptr, _absmax / fp8_max)
+
+
+@triton.jit
+def _per_tensor_quant_mla_fp8_stage2(
+    x_ptr,
+    x_s_ptr,
+    x_q_ptr,
+    num_seq,
+    head_size,
+    x_stride_h,
+    x_stride_s,
+    fp8_min,
+    fp8_max,
+    BLOCK_SIZE: tl.constexpr,
+):
+    seq_id = tl.program_id(0)
+    head_id = tl.program_id(1)
+    offset = tl.arange(0, BLOCK_SIZE)
+    mask = offset < head_size
+
+    x_s = tl.load(x_s_ptr)
+    x_s_inv = 1.0 / x_s
+
+    x_ptr += head_id * x_stride_h + seq_id * x_stride_s
+    x_q_ptr += head_id * num_seq * head_size + seq_id * head_size
 
     x = tl.load(x_ptr + offset, mask=mask, other=0.0).to(tl.float32)
     x_q = tl.clamp(x * x_s_inv, fp8_min, fp8_max).to(x_q_ptr.dtype.element_ty)
     tl.store(x_q_ptr + offset, x_q, mask=mask)
-
-    if seq_id == 0 and head_id == 0:
-        tl.store(x_s_ptr, 1.0 / x_s_inv)
 
 
 def per_tensor_quant_mla_fp8(
@@ -960,24 +909,30 @@ def per_tensor_quant_mla_fp8(
         fp8_max = 224.0
 
     x_q = x.new_empty(x.size(), dtype=dtype)
-    x_s = x.new_empty((1,), dtype=torch.float32)
+    x_s = torch.zeros((1,), dtype=torch.float32, device=x.device)
 
     num_head, num_seq, head_size = x.shape
     BLOCK_SIZE = triton.next_power_of_2(head_size)
     grid = (num_seq, num_head)
 
-    min_val, max_val = x.aminmax()
-    _per_tensor_quant_mla_fp8[grid](
+    _per_tensor_quant_mla_fp8_stage1[grid](
         x,
-        min_val,
-        max_val,
+        x_s,
+        head_size,
+        x.stride(0),
+        x.stride(1),
+        eps,
+        fp8_max,
+        BLOCK_SIZE,
+    )
+    _per_tensor_quant_mla_fp8_stage2[grid](
+        x,
         x_s,
         x_q,
         num_seq,
         head_size,
         x.stride(0),
         x.stride(1),
-        eps,
         -fp8_max,
         fp8_max,
         BLOCK_SIZE,