<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation

1. fix `input_to_float8` precision
2. impl mla absrob specialized triton per tensor quant fp8
3. optimize out Tensor.aminmax which call Tensor.contiguous inside

<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Unittest

```bash
python3 python/sglang/test/test_block_fp8.py            
test_per_tensor_quant_mla_fp8 (__main__.TestPerTensorQuantMlaFP8) ... ok
```

## Perf

About 22% boost🚀 on decode phase

### 1-batch

cmd: `python3 -m sglang.bench_one_batch --model RedHatAI/DeepSeek-Coder-V2-Lite-Instruct-FP8 --batch-size 1 --input-len 128 --output-len 128 --trust-remote-code`

- before
```
Benchmark ...
Prefill. latency: 0.07211 s, throughput:   1775.16 token/s
Decode.  latency: 0.00735 s, throughput:    136.03 token/s
Decode.  latency: 0.00735 s, throughput:    136.01 token/s
Decode.  latency: 0.00736 s, throughput:    135.86 token/s
Decode.  latency: 0.00732 s, throughput:    136.53 token/s
Decode.  latency: 0.00734 s, throughput:    136.17 token/s
Decode.  median latency: 0.00732 s, median throughput:    136.69 token/s
Total. latency:  1.003 s, throughput:    255.15 token/s
```

- after
```
Benchmark ...
Prefill. latency: 0.06900 s, throughput:   1855.09 token/s
Decode.  latency: 0.00632 s, throughput:    158.16 token/s
Decode.  latency: 0.00633 s, throughput:    157.97 token/s
Decode.  latency: 0.00631 s, throughput:    158.36 token/s
Decode.  latency: 0.00629 s, throughput:    159.01 token/s
Decode.  latency: 0.00632 s, throughput:    158.20 token/s
Decode.  median latency: 0.00628 s, median throughput:    159.35 token/s
Total. latency:  0.868 s, throughput:    295.07 token/s
```

### 64-batch

cmd: `python3 -m sglang.bench_one_batch --model RedHatAI/DeepSeek-Coder-V2-Lite-Instruct-FP8 --batch-size 64 --input-len 128 --output-len 128 --trust-remote-code`

- before
```
Benchmark ...
Prefill. latency: 0.10302 s, throughput:  79520.42 token/s
Decode.  latency: 0.00911 s, throughput:   7026.19 token/s
Decode.  latency: 0.00910 s, throughput:   7032.81 token/s
Decode.  latency: 0.00909 s, throughput:   7040.19 token/s
Decode.  latency: 0.00909 s, throughput:   7043.70 token/s
Decode.  latency: 0.00911 s, throughput:   7028.76 token/s
Decode.  median latency: 0.00905 s, median throughput:   7071.16 token/s
Total. latency:  1.253 s, throughput:  13076.68 token/s
```

- after
```
Benchmark ...
Prefill. latency: 0.08751 s, throughput:  93607.45 token/s
Decode.  latency: 0.00743 s, throughput:   8608.94 token/s
Decode.  latency: 0.00744 s, throughput:   8605.91 token/s
Decode.  latency: 0.00739 s, throughput:   8655.86 token/s
Decode.  latency: 0.00741 s, throughput:   8639.98 token/s
Decode.  latency: 0.00739 s, throughput:   8663.40 token/s
Decode.  median latency: 0.00738 s, median throughput:   8677.96 token/s
Total. latency:  1.024 s, throughput:  15996.97 token/s
```

## Acc

cmd: `python3 test_mla_deepseek_v3.py`

- before: 0.645~0.665
- after: 0.64~0.67

## Modifications

<!-- Describe the changes made in this PR. -->

## Checklist

- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [x] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [x] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.