@Alcanderian 2025-04-14
> Can you add gsm8k acc test too?

Ok, I will post the accuracy report by TestMLADeepseekV3 at https://github.com/sgl-project/sglang/blob/main/test/srt/test_mla_deepseek_v3.py#L17 after I fully finish this PR