@Alcanderian 2025-04-14
> Can you add gsm8k acc test too?

Done. The failed cases are not related to deepseek fp8.

## Acc

cmd: `python3 test_mla_deepseek_v3.py`

- before: 0.645~0.665
- after: 0.665~0.68