@zhyncs 2025-04-14
Three gold bro is GOAT!