@BBuf 2025-04-15
> `_per_tensor_quant_fp8` has been removed and `_per_tensor_quant_mla_fp8` has been optimized(boost 64-batch case from 8000token/s to 8600 token/s). Please take a look again XD @BBuf

Great job, I have no other advice.